layout: Move the `LayoutNode` wrapper from script into layout.
