+++ src/components/main/css/node_style.rs
@@ -6,10 +6,10 @@
 
 use css::node_util::NodeUtil;
 use layout::incremental::RestyleDamage;
+use layout::wrapper::LayoutNode;
 
 use extra::arc::Arc;
 use style::ComputedValues;
-use script::dom::node::LayoutNode;
 
 /// Node mixin providing `style` method that returns a `NodeStyle`
 pub trait StyledNode {
