+++ src/components/main/servo.rc
@@ -88,6 +88,7 @@ pub mod layout {
     pub mod text;
     pub mod util;
     pub mod incremental;
+    pub mod wrapper;
     mod extra;
 }
